Changes to packaging for Analyer Utilities package
1. Rename `Microsoft.CodeAnalysis.FlowAnalysis.Utilities` to `Microsoft.CodeAnalysis.AnalyzerUtilities`: This package now contains following utilities:
   1. Flow analysis utilities
   2. Code metrics utilities
   3. RulesetToEditorconfigConverter utility
2. Rename NuGet package `RulesetToEditorconfigConverter` to `Microsoft.CodeAnalysis.RulesetToEditorconfigConverter`: This matches the naming prefix for all of our published NuGet packages in this repo. Additionally, this package now only contains the executable for converting ruleset to editorconfig file. The library for the same functionality is part of `Microsoft.CodeAnalysis.AnalyzerUtilities`.
