+++ nuget/Microsoft.CodeAnalysis.AnalyzerUtilities/Microsoft.CodeAnalysis.AnalyzerUtilities.Package.csproj
@@ -1,0 +1,34 @@
+﻿<!-- Copyright (c)  Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information. -->
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <IsPackable>true</IsPackable>
+    <IncludeBuildOutput>false</IncludeBuildOutput>
+    <NuspecPackageId>Microsoft.CodeAnalysis.AnalyzerUtilities</NuspecPackageId>
+    <Description>Analyzer utilities for various analyses, including Dataflow analysis based on ControlFlowGraph API in Microsoft.CodeAnalysis.</Description>
+    <Summary>Analyzer utilities for various analyses, including Dataflow analysis</Summary>
+    <ReleaseNotes>Analyzer utilities for various analyses, including Dataflow analysis based on ControlFlowGraph API in Microsoft.CodeAnalysis.</ReleaseNotes>
+    <PackageTags>Roslyn Analyzer Utilities CodeAnalysis Dataflow ControlFlowGraph FlowAnalysis Compiler CSharp VB VisualBasic Diagnostic Analyzers Syntax Semantics</PackageTags>
+
+    <!-- 
+      PackageId is used by Restore. If we set it to Microsoft.CodeAnalysis.AnalyzerUtilities
+      Restore would conclude that there is a cyclic dependency between Microsoft.CodeAnalysis and Microsoft.CodeAnalysis.AnalyzerUtilities.
+    -->
+    <PackageId>*$(MSBuildProjectFullPath)*</PackageId>
+    <DevelopmentDependency>false</DevelopmentDependency>
+
+    <!-- Override the version prefix for this package to allow more frequent stable releases -->
+    <VersionPrefix>$(AnalyzerUtilitiesVersionPrefix)</VersionPrefix>
+
+    <NoWarn>$(NoWarn);NU5128</NoWarn>
+  </PropertyGroup>
+  
+  <ItemGroup>
+    <AnalyzerNupkgLibrary Include="Microsoft.CodeAnalysis.AnalyzerUtilities.dll" />
+    <AnalyzerNupkgLibrary Include="Microsoft.CodeAnalysis.AnalyzerUtilities.xml" />
+  </ItemGroup>
+  
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Microsoft.CodeAnalysis.AnalyzerUtilities\Microsoft.CodeAnalysis.AnalyzerUtilities.csproj" />
+  </ItemGroup>
+</Project>
