--- nuget/Microsoft.CodeAnalysis.FlowAnalysis.Utilities/Microsoft.CodeAnalysis.FlowAnalysis.Utilities.Package.csproj
@@ -1,32 +1,0 @@
-﻿<!-- Copyright (c)  Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information. -->
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <IsPackable>true</IsPackable>
-    <IncludeBuildOutput>false</IncludeBuildOutput>
-    <NuspecPackageId>Microsoft.CodeAnalysis.FlowAnalysis.Utilities</NuspecPackageId>
-    <Description>Utilities for Dataflow analysis based on ControlFlowGraph API in Microsoft.CodeAnalysis.</Description>
-    <Summary>Utilities for Dataflow analysis</Summary>
-    <ReleaseNotes>Utilities for Dataflow analysis  based on ControlFlowGraph API in Microsoft.CodeAnalysis.</ReleaseNotes>
-    <PackageTags>Roslyn CodeAnalysis Dataflow ControlFlowGraph FlowAnalysis Compiler CSharp VB VisualBasic Diagnostic Analyzers Syntax Semantics</PackageTags>
-
-    <!-- 
-      PackageId is used by Restore. If we set it to Microsoft.CodeAnalysis.FlowAnalysis.Utilities
-      Restore would conclude that there is a cyclic dependency between Microsoft.CodeAnalysis and Microsoft.CodeAnalysis.FlowAnalysis.Utilities.
-    -->
-    <PackageId>*$(MSBuildProjectFullPath)*</PackageId>
-    <DevelopmentDependency>false</DevelopmentDependency>
-
-    <!-- Override the version prefix for this package to allow more frequent stable releases -->
-    <VersionPrefix>$(FlowAnalysisUtilitiesVersionPrefix)</VersionPrefix>
-  </PropertyGroup>
-  
-  <ItemGroup>
-    <AnalyzerNupkgLibrary Include="Microsoft.CodeAnalysis.FlowAnalysis.Utilities.dll" />
-    <AnalyzerNupkgLibrary Include="Microsoft.CodeAnalysis.FlowAnalysis.Utilities.xml" />
-  </ItemGroup>
-  
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\Microsoft.CodeAnalysis.FlowAnalysis.Utilities\Microsoft.CodeAnalysis.FlowAnalysis.Utilities.csproj" />
-  </ItemGroup>
-</Project>
