+++ nuget/Microsoft.CodeAnalysis.RulesetToEditorconfigConverter/Microsoft.CodeAnalysis.RulesetToEditorconfigConverter.Package.csproj
@@ -1,11 +1,11 @@
 ﻿<!-- Copyright (c)  Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information. -->
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>net472;netstandard2.0</TargetFrameworks>
+    <TargetFrameworks>net472</TargetFrameworks>
     
     <IsPackable>true</IsPackable>
     <IncludeBuildOutput>false</IncludeBuildOutput>
-    <NuspecPackageId>RulesetToEditorconfigConverter</NuspecPackageId>
+    <NuspecPackageId>Microsoft.CodeAnalysis.RulesetToEditorconfigConverter</NuspecPackageId>
     <Description>Utility to convert ruleset files to equivalent .editorconfig files. Editorconfig files are respected by C# and VB compilers on VS2019 16.3 or later. See https://docs.microsoft.com/visualstudio/code-quality/use-roslyn-analyzers#set-rule-severity-in-an-editorconfig-file for details.</Description>
     <Summary>Utility to convert ruleset files to equivalent .editorconfig files</Summary>
     <ReleaseNotes>Utility to convert ruleset files to equivalent .editorconfig files</ReleaseNotes>
@@ -13,13 +13,11 @@
     <ContainsPortedFxCopRules>false</ContainsPortedFxCopRules>
 	  <GeneratePackagePropsFile>false</GeneratePackagePropsFile>
     
-	  <NoWarn>$(NoWarn);NU5100;NU5128</NoWarn>
+	  <NoWarn>$(NoWarn);NU5100</NoWarn>
   </PropertyGroup>
   
   <ItemGroup>
     <AnalyzerNupkgFolder Include="RulesetToEditorconfigConverter" />
-    <AnalyzerNupkgLibrary Include="RulesetToEditorconfigConverter.dll" />
-    <AnalyzerNupkgLibrary Include="RulesetToEditorconfigConverter.xml" />
   </ItemGroup>
   
   <ItemGroup>
