+++ src/Microsoft.CodeAnalysis.AnalyzerUtilities/Microsoft.CodeAnalysis.AnalyzerUtilities.csproj
@@ -3,14 +3,14 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
     <!--
-      PackageId is used by Restore. If we set it to DotNetAnalyzers.BannedApiAnalyzer,
-      Restore would conclude that there is a cyclic dependency between us and the DotNetAnalyzers.BannedApiAnalyzer package.
+      PackageId is used by Restore. If we set it to Microsoft.CodeAnalysis.AnalyzerUtilities,
+      Restore would conclude that there is a cyclic dependency between us and the Microsoft.CodeAnalysis.AnalyzerUtilities package.
     -->
     <PackageId>*$(MSBuildProjectFullPath)*</PackageId>
     <ExcludeInternalFlowAnalyses>true</ExcludeInternalFlowAnalyses>
     <ExcludeCodeMetricsUtilities>true</ExcludeCodeMetricsUtilities>
 
-    <!-- RS0026: Avoud public API overloads with differences in optional parameters -->
+    <!-- RS0026: Avoid public API overloads with differences in optional parameters -->
     <NoWarn>$(NoWarn);RS0026</NoWarn>
   </PropertyGroup>
   <Import Project="..\Utilities\Compiler\Analyzer.Utilities.projitems" Label="Shared" />
