+++ src/Utilities/Compiler/Extensions/CompilationExtensions.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
 using Microsoft.CodeAnalysis;
 
 namespace Analyzer.Utilities.Extensions
